backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
langchain
langchain-core
langchain-community
langchain-openai
langchain-groq
langchain-chroma
chromadb
sentence-transformers # embeddings to run locally
textblob
textblob
pytest
httpx
python-jose[cryptography]
passlib[bcrypt]
python-multipart                